Utilities for `*_cert` in `common_cert.go`, and utilities for `*_key` in `common_key.go`
