+++ internal/provider/util.go
@@ -1,12 +1,7 @@
 package provider
 
 import (
-	"crypto"
-	"crypto/ecdsa"
-	"crypto/ed25519"
-	"crypto/rsa"
 	"crypto/sha1"
-	"crypto/x509"
 	"encoding/hex"
 	"encoding/pem"
 	"fmt"
@@ -14,7 +9,6 @@ import (
 	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"golang.org/x/crypto/ssh"
 )
 
 func decodePEM(d *schema.ResourceData, pemKey, pemType string) (*pem.Block, error) {
@@ -27,137 +21,6 @@ func decodePEM(d *schema.ResourceData, pemKey, pemType string) (*pem.Block, erro
 	}
 
 	return block, nil
-}
-
-func parsePrivateKey(d *schema.ResourceData, pemKey, algoKey string) (crypto.PrivateKey, error) {
-	algoName := Algorithm(d.Get(algoKey).(string))
-
-	parser, ok := keyParsers[algoName]
-	if !ok {
-		return nil, fmt.Errorf("invalid %s: %#v", algoKey, algoName)
-	}
-
-	block, err := decodePEM(d, pemKey, "")
-	if err != nil {
-		return nil, err
-	}
-
-	key, err := parser(block.Bytes)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode %s: %s", pemKey, err)
-	}
-
-	return key, nil
-}
-
-func parseCertificate(d *schema.ResourceData, pemKey string) (*x509.Certificate, error) {
-	block, err := decodePEM(d, pemKey, "")
-	if err != nil {
-		return nil, err
-	}
-
-	certs, err := x509.ParseCertificates(block.Bytes)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse %s: %s", pemKey, err)
-	}
-	if len(certs) < 1 {
-		return nil, fmt.Errorf("no certificates found in %s", pemKey)
-	}
-	if len(certs) > 1 {
-		return nil, fmt.Errorf("multiple certificates found in %s", pemKey)
-	}
-
-	return certs[0], nil
-}
-
-func parseCertificateRequest(d *schema.ResourceData, pemKey string) (*x509.CertificateRequest, error) {
-	block, err := decodePEM(d, pemKey, CertificateRequest.String())
-	if err != nil {
-		return nil, err
-	}
-
-	certReq, err := x509.ParseCertificateRequest(block.Bytes)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse %s: %s", pemKey, err)
-	}
-
-	return certReq, nil
-}
-
-func certificateToMap(cert *x509.Certificate) map[string]interface{} {
-	return map[string]interface{}{
-		"signature_algorithm":  cert.SignatureAlgorithm.String(),
-		"public_key_algorithm": cert.PublicKeyAlgorithm.String(),
-		"serial_number":        cert.SerialNumber.String(),
-		"is_ca":                cert.IsCA,
-		"version":              cert.Version,
-		"issuer":               cert.Issuer.String(),
-		"subject":              cert.Subject.String(),
-		"not_before":           cert.NotBefore.Format(time.RFC3339),
-		"not_after":            cert.NotAfter.Format(time.RFC3339),
-		"sha1_fingerprint":     fmt.Sprintf("%x", sha1.Sum(cert.Raw)),
-	}
-}
-
-// setPublicKeyAttributes takes a crypto.PrivateKey, extracts the corresponding crypto.PublicKey and then
-// encodes related attributes on the given schema.ResourceData.
-func setPublicKeyAttributes(d *schema.ResourceData, prvKey crypto.PrivateKey) error {
-	pubKey := toPublicKey(prvKey)
-	pubKeyBytes, err := x509.MarshalPKIXPublicKey(pubKey)
-	if err != nil {
-		return fmt.Errorf("failed to marshal public key error: %s", err)
-	}
-	pubKeyPemBlock := &pem.Block{
-		Type:  PublicKey.String(),
-		Bytes: pubKeyBytes,
-	}
-
-	d.SetId(hashForState(string(pubKeyBytes)))
-
-	if err := d.Set("public_key_pem", string(pem.EncodeToMemory(pubKeyPemBlock))); err != nil {
-		return fmt.Errorf("error setting value on key 'public_key_pem': %s", err)
-	}
-
-	// NOTE: ECDSA keys with elliptic curve P-224 are not supported by `x/crypto/ssh`,
-	// so this will return an error: in that case, we set the below fields to emptry strings
-	sshPubKey, err := ssh.NewPublicKey(toPublicKey(prvKey))
-	var pubKeySSH, pubKeySSHFingerprintMD5, pubKeySSHFingerprintSHA256 string
-	if err == nil {
-		sshPubKeyBytes := ssh.MarshalAuthorizedKey(sshPubKey)
-
-		pubKeySSH = string(sshPubKeyBytes)
-		pubKeySSHFingerprintMD5 = ssh.FingerprintLegacyMD5(sshPubKey)
-		pubKeySSHFingerprintSHA256 = ssh.FingerprintSHA256(sshPubKey)
-	}
-
-	if err := d.Set("public_key_openssh", pubKeySSH); err != nil {
-		return fmt.Errorf("error setting value on key 'public_key_openssh': %s", err)
-	}
-
-	if err := d.Set("public_key_fingerprint_md5", pubKeySSHFingerprintMD5); err != nil {
-		return fmt.Errorf("error setting value on key 'public_key_fingerprint_md5': %s", err)
-	}
-
-	if err := d.Set("public_key_fingerprint_sha256", pubKeySSHFingerprintSHA256); err != nil {
-		return fmt.Errorf("error setting value on key 'public_key_fingerprint_sha256': %s", err)
-	}
-
-	return nil
-}
-
-// toPublicKey takes a crypto.PrivateKey and extracts the corresponding crypto.PublicKey,
-// after having figured out its type.
-func toPublicKey(prvKey crypto.PrivateKey) crypto.PublicKey {
-	switch k := prvKey.(type) {
-	case *rsa.PrivateKey:
-		return &k.PublicKey
-	case *ecdsa.PrivateKey:
-		return &k.PublicKey
-	case *ed25519.PrivateKey:
-		return k.Public()
-	default:
-		return nil
-	}
 }
 
 // hashForState computes the hexadecimal representation of the SHA1 checksum of a string.
